chore: Update getUserAlbums import in secure/page.tsx
