+++ my-app/src/app/ev1/page.tsx
@@ -4,6 +4,7 @@ import { getDatabase, ref, child, get } from "firebase/database";
 import { initializeApp } from 'firebase/app';
 import { firebaseConfig } from "@/app/firebase/firebaseConfig"
 import { writeNewAlbum, writeNewPicture } from "../lib/write_functions_rtdb";
+import { getUserAlbums } from "./functions";
 
 const firebaseApp = initializeApp(firebaseConfig);
 
@@ -58,21 +59,7 @@ export function getUserAlbums(userName) {
 }
  */
 
-export async function getUserAlbums(userName) {
-    let unr = userName.replace(/[\.\#$\[\]@]/g, '').trim();
-    try {
-        const databaseRef = ref(getDatabase());
-        const snapshot = await get(child(databaseRef, `/album/${unr}`));
-        if (snapshot.exists()) {
-            const albumsObject = snapshot.val();
-            return albumsObject;
-        } else {
-            throw new Error("No data found");
-        }
-    } catch (error) {
-        throw new Error(`Failed because of: ${error}`);
-    }
-}
+
 
 type StorageType = {
     imageUrl: string,
